增加check_info_confirm 云函数
