+++ cloudfunctions/check_Info_Confirm/config.json
@@ -1,0 +1,6 @@
+{
+  "permissions": {
+    "openapi": [
+    ]
+  }
+}
