+++ cloudfunctions/check_Info_Confirm/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "checkInfoConfirm",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "wx-server-sdk": "~2.0.4"
+  }
+}
